Remove unpredictable tests (2x bad input in same request)
Removing tests that set bad IN and EG RL in same request and expect
error message to report 1 particular error out of 2.

In openstack master (Rocky onwards) apparently the order changes for
which error gets called first, must be some upstream change changing
order of how params are presented to the code.

Rather than extending the test for one case or the other depending
on the OS version, or just find a subnet of the regex string to match
always, i judge this is just bad test design and can be omitted.

Other tests test specifically for bad IN or EG RL.

Change-Id: I7a3ac5fe403e23246467de1cf4ce602d0052bbb5