Connectivity tests further refactored
The l2 connectivity approach is further abstracted towards the test
engineer now. The 'create_tenant_server' method takes a 'make_reachable'
argument and if set, depending on the network that server is deployed
in is l2 or l3, the needed actions are taken. In order to find out l2 vs
l3, extra metadata is added to the network and subnet dicts.

The actual approach of how a l2 topology is transformed to l3 remains
unchanged.

Also a server's dualstack itf configuration is now handled by
'create_tenant_server' based on 'configure_dualstack_itf' new added
parameter.

Change-Id: I99c99e4634de8968525171418a934766f6350403