Reserve FIP address in OS for vsd managed FIP
Currently, VSD managed workflows in connectivity tests
will allocate FIP's and assign them to a vms. Openstack
doesn't know about those FIP allocations, and if OS managed
tests are run in parallel, there is a possibility of a conflict,
resulting in test failures.
This patch set will effectively allocate FIP in openstack, and use
this address for FIP allocation in VSD to prevent such a conflicts.

Change-Id: I8fb2bf9c3a9d91ce2644aedf2695512a35db19b4
Depends-On: I54b80b5a5ef264f03981b6fc5174eb6a41d0fad4