Do not fail tests on adding default ipv6 route
Today, we have a sporadic failures in dualstack connectivity
tests when adding default ipv6 route to a VM. Error seen in
this case suggests that default route is already present.
It is unclear at this point why it does happen, but it is not
a good reason to fail test. This patchset makes sure we ignore
such errors.

Change-Id: I7e090af0394a4808bd253dcb7b862bdcb2b04519
Depends-On: I54b80b5a5ef264f03981b6fc5174eb6a41d0fad4