Fix default netpartition access
Fixing I2d66c6310a226a341a0942a8b0324dcf0f74513d

This commit fixes the overloaded way how 
BaseVSDManagedNetworksIPv6Test overloads the NuageBaseTest
net_partition field, which in the base class is just the name
of the netpartition (string) while is overloaded to a 
list of objects. This caused the mismatch in how it is accessed
in the indivual tests. This is now cleaned up.

Change-Id: Ib52d5c99c4820a0ac3afce5c045ab80ad1a469ba