Further rationalize connectivity tests
This patchset addresses 3 main goals

1) Rationalize use of 'prepare_for_connectivity'

2) Rationalize inconsistent way of client being passed
   to nuage_test baseline methods, while
   a) what is passed is a manager
   b) not being propagated all the way, in many cases

3) Propagating use of cleanup flag for all resources fully
   now, and supporting no-cleanup at every level

---

1) Rationalize use of 'prepare_for_connectivity'

The reason why servers have a 'prepare_for_connectivity'
flag is because when a server is deployed in l2 network
and we know it will need FIP access,
we want to give it a nic in a new l3 network, at time of
boot, as to avoid hotplugging complexity and so-forth.

Also, while that was originally not the case, people preferred
a FIP to be allocated at time of server create, and not
at the time this would turn out to be needed.

Also, and this is important to remember,
'prepare_for_connectivity' is being set in order to allow
interrogating the server whether cloudinit successfully
completed. This may go against the expected business logic
where the target server in a assert_ping operation wouldn't
need to be prepared for connectivity (no need for a FIP),
interroging that server for having completed its cloudinit
is always safer than relying on hidden assumptions, and which
to this patchset's author's opinion could be at the origin of
the connectivity test instabilities repeatedly seen in CI:
- given i have spawn my target server before my jump server,
  and given my jump server did now complete its cloudinit,
  i can safely assume the target server will as well
- given the target's server topology is simpler and given
  the above... (etc)

Hence, in order to, a)
   remove from business logic, preparing a server for
   connectivity while from a logical perspective it wouldn't
and, b)
   still make connectivity tests most robust for any type
   of deployment case (slow/fast/vm/bm...),
proposed is a default_prepare_for_connectivity flag, which
defaults to False at base class level but which scenario
tests can/will set to True to cover robust cloudinit checks.

By convention, the jump server is still created with
prepare_for_connectivity set. Other cases where
'prepare_for_connectivity' is still set, by convention, are:
- you will run your custom scripts on that server, e.g. for
  setting it up as a web server
- the server needs provisioning,
  e.g. when the server is deployed in dualstack network, or
       when the server needs dot1q interface(s)

Other improvements are:

- A timestamp is set in the /tmp/cloudinit_completed file
  and on checking the completion, an entra elapse time of 3
  secs is guaranteed. If cloudinit completed more than 3 secs
  ago, no wait time is added.
- The 'complete_prepare_for_connectivity' method disappears.
  The need to provision a server is tracked by an internal
  server attribute needs_provisioning, checked at assert_ping.
  The method taking the action is provision().
  NOTE : this method is meant to disappear, when everything is
  done thru cloudinit. TOOD(Kris) in subsequent patchset.
- A on-the-server sanity dhcp-check when was implemented at the time
  that connectivity were highly instable as of single ipv6.
  That is now removed.

----

2) Rationalize propagation of manager being passed

An attribute 'manager' is now passed when not the client itself (e.g.
networks_client) but the manager of that client is being passed.

----

3) Full propagation of cleanup everywhere

'cleanup' flag is now propogated fully.

Also, as upstream _create_security_group method does not support
'cleanup', it is copied and added now. That is a small upstream
contribution we could do.

----

Change-Id: Icab30200c72012e9c716dce70be0a0dfcbca3385