Fix test_icmp_connectivity_multiple_subnets_in_shared_network
Intention of test was to check that two subnets of an external
network end up in the same L3 domain on VSD and that there
is connectivity between them.

The test mistakenly used underlay=True subnets that where
not reachable on real deployments.

Fix is to have server 1 with a nic in the underlay
external and other nic in non-underlay external.
Server 2 only has a nic in non-underlay external.

Change-Id: I9ccc48cfe2a54a080ab879e8bea2867101a9cb12