Delete original gash/QA IpAntiSpoofingTestScenario test
This is one of the tests from the very early days when tempest
tests were run out of gash and had a hook mechanism to check
stuff on VRS. While all the idea's of that were good, it never
flew really. This is a test that i never further paid attention
to when doing the original port from gash/QA to upstream tempest
way of doing things.

This patchset is cleaning up that (dead) code now.

Change-Id: If05097e66c875e3d369ff100f47129b8bc872e6b