Give clearer error when v6 IP remains tentative
Difference with v4 and v6 is that a v6 address first gets established
as tentative IP. Then the DAD process takes place and when good it
becomes permanent.
This patchset adds clarity in the test failure reporting, that when a
v6 ip does not establish as permanent, it is
(a) as it does not get established at all, or
(b) it does but remains tentative (as DAD failed, likely).
Today that difference is not made.

Change-Id: Id837f28418c096e57a012f09eaff319d88913ed7