Clean name randomize support
Introduce 2 methods for reusable name randomization support
(one classmethod and one non-classmethod).

For names obtained in class methods, the randomized name will hold
the class name (or part of it), for traceability.

For names obtained in (non-class method) test methods, it will hold
the test name (or part of it), for traceability.

That makes it easier post-test execution to see where leftover objects
originate from.

Depends-On: http://gerrit.an.nuagenetworks.net:8080/c/OpenStack/nuage-tempest-plugin/+/176
Change-Id: Ifccf176968cc3a9907b2384204ce8837214a9718