Replace cloudinit completion check with console-log
Instead of SSH'ing into the instance and check a file
written by user data, check the console log for a tag
to appear, echo'ed from user data.

Note: the last 50 lines of the console log are retrieved
per iteration, only.

Advantages are:
1. more robustness implementation on slow systems
2. not every instance needs a FIP

The latter item will be further anticipated to in a
next patchset.

Change-Id: I6317fb554cdbc9fe1e5904ae555a2371db3be92d